Fix physics profile parsing: add error logging, make types forgiving
The extracted JSON was being silently rejected by Codable. Fixes:
- Add decode error logging (prints exact error + available keys)
- QuarkMechanism.type now Optional (model sometimes omits)
- RhythmData waveforms now [Double] (model outputs floats not ints)
- All .type references updated with ?? "" or optional chaining
- Copy function uses compactMap for optional types

Data is NOT lost — it's saved in Supabase. This fix ensures the
Swift Codable parser can handle the model's actual JSON output.

+++ UI/FocusMode/SwipeStudy/PhysicsRhythmView.swift
@@ -32,13 +32,13 @@ struct PhysicsRhythmView: View {
     private func densityBars(_ r: RhythmData) -> some View {
         let density = r.densityWaveform ?? []
         let energy = r.energyCurve ?? []
-        let maxDensity = max(Double(density.max() ?? 1), 1)
+        let maxDensity = max(density.max() ?? 1.0, 1.0)
 
         VStack(alignment: .leading, spacing: DS.space4) {
             HStack(alignment: .bottom, spacing: 2) {
                 ForEach(Array(density.enumerated()), id: \.offset) { index, wordCount in
-                    let height = max(CGFloat(Double(wordCount) / maxDensity) * 40, 4)
-                    let energyLevel = index < energy.count ? energy[index] : 3
+                    let height = max(CGFloat(wordCount / maxDensity) * 40, 4)
+                    let energyLevel = index < energy.count ? Int(energy[index]) : 3
                     let silence = (r.silenceSlides ?? []).contains(index + 1)
 
                     VStack(spacing: 1) {
